# [crypto] parallelize witness commitment and interpolation